requirements.txt
streamlit==1.18.1
qdrant-client
langchain==0.0.184
PyPDF2==3.0.1
python-dotenv==1.0.0
openai==0.27.6
faiss-cpu==1.7.4
altair==4
tiktoken==0.4.0
huggingface-hub==0.14.1
InstructorEmbedding==1.0.1
sentence-transformers==2.2.2
pypdf
torch
accelerate
bitsandbytes
transformers
sentence_transformers
chainlit


